+++ veld-processor/src/main/java/module-info.java
@@ -2,21 +2,21 @@
  * Veld Processor Module.
  * Annotation processor that generates bytecode using ASM.
  */
-module com.veld.processor {
+module io.github.yasmramos.processor {
     // Required JDK modules
     requires java.compiler;
     
     // Required Veld modules
-    requires com.veld.annotation;
-    requires com.veld.runtime;
+    requires io.github.yasmramos.annotation;
+    requires io.github.yasmramos.runtime;
     
     // ASM modules for bytecode generation
     requires org.objectweb.asm;
     
     // Export processor package
-    exports com.veld.processor;
+    exports io.github.yasmramos.processor;
     
     // Provide annotation processor service
     provides javax.annotation.processing.Processor 
-        with com.veld.processor.VeldProcessor;
+        with io.github.yasmramos.processor.VeldProcessor;
 }